requirements.txt
python-dotenv
langchain
langchain[groq]
pydantic
langchain_community
langchain_core
#wikepedia
